将最大值改为Double型的最大值

+++ source/silhouette/src/Silhouette.java
@@ -34,7 +34,7 @@ class Silhouette{
 
     double getDistanceWithOtherCluster(int instanceNumber, int clusterNumber){
 
-        double minDistance = Integer.MAX_VALUE;
+        double minDistance = Double.MAX_VALUE;
         for (int i=0; i<clusters.length; i++){
             if ( i != clusterNumber){
                 double[] cluster = clusters[i];
